renamed participant.landing test to match the route file name
